+++ special-pages/pages/history/app/components/Sidebar.js
@@ -6,8 +6,7 @@ import { useTypedTranslation } from '../types.js';
 import { Trash } from '../icons/Trash.js';
 import { useTypedTranslationWith } from '../../../new-tab/app/types.js';
 import { useQueryContext, useQueryDispatch } from '../global/Providers/QueryProvider.js';
-import { useResultsData } from '../global/Providers/DataProvider.js';
-import { useHistoryServiceDispatch } from '../global/Providers/HistoryServiceProvider.js';
+import { useHistoryServiceDispatch, useResultsData } from '../global/Providers/HistoryServiceProvider.js';
 
 /**
  * @import json from "../strings.json"
